Base sync up to 969bf50 (#432-#436, #414/#415/#423).

Conflicts:
- app/CMakeLists.txt: build both app_radio.c and app_slot.c.
- app/src/main.c: nfc_result_before_reboot() (#414) plus LOG_WRN_REBOOTING
  and the P2P-only CONFIG_LORAWAN guards.
- app/src/app_report.c: keep the RTC slot cadence (#436) and route through
  app_radio (no CONFIG_LORAWAN guards); force path via the new
  app_radio_send_telemetry_now() (F14; P2P has no pre-send jitter).
- app/src/app_cmd.c: keep radio_transport() and the host-driven NFC pages.
- tests/cmd: keep the P2P paging test and the #433 slot-ROM test.
- tests/nfc_hw: take v1.5.0 (mailbox rewrite supersedes the reformat).
- doc/version 1.5.md: P2P stays §6, v1.5.0 §14-§20 renumbered to §15-§21,
  GetSettings transports incl. P2P; cross-doc refs updated.

Fixes in auto-merged hunks:
- app_cmd.c dispatch regenerated: the merged region dropped p2p from every
  default-transport command (get_info/get_config/set_param/... would answer
  NOT_READY over P2P).
- get_config skips the dump_lrw:false slot ROMs on P2P too; page 0 and the
  LRW-laid-out stream pages disagreed on page_count.
- app_p2p.c history replay ported to the #436 API: absolute cursor via
  app_history_span()/export_abs(), per-frame time_synced, ends at the
  snapshot end; p2p_logic stubs updated, cursor test added.
- 'ats lrw status' in new v1.5.0 doc text -> 'ats radio status'.

+++ app/src/app_p2p.h
@@ -402,7 +402,7 @@ void p2p_test_get_join(uint8_t *sf, uint8_t *step, uint8_t *attempts, bool *slow
 void p2p_test_set_join_started_at(int64_t at_ms);
 struct p2p_duty *p2p_test_get_duty(void);
 void p2p_test_set_replay_active(bool active);
-void p2p_test_get_replay(bool *active, uint32_t *seq, size_t *cursor, uint32_t *idx);
+void p2p_test_get_replay(bool *active, uint32_t *seq, uint32_t *cursor, uint32_t *idx);
 void p2p_test_build_join_request(uint32_t dev_nonce, uint8_t out[P2P_JOIN_REQ_LEN]);
 void p2p_test_derive_session_key(uint32_t dev_nonce, uint32_t central_nonce,
 				 uint8_t out[P2P_KEY_LEN]);
